Added some features for next extension of the project(Connecting tg bot for reminding). New fresh design is also implemented

+++ todo/tasks/views.py
@@ -10,17 +10,18 @@ def home(request):
 def todo(request):
     if request.method == 'POST':
         title = request.POST.get('title')
-        Task.objects.create(title=title, user=request.user)
+        deadline = request.POST.get('deadline')
+        Task.objects.create(title=title, deadline=deadline, user=request.user)
         return redirect('/todopage/')
 
-    # Запрашиваем задачи текущего пользователя
     todos = Task.objects.filter(user=request.user).order_by('-date')
     return render(request, 'todo.html', {'res': todos})
 
 @login_required(login_url='/loginn/')
 def delete_todo(request, srno):
-    obj = get_object_or_404(Task, srno=srno, user=request.user)
-    obj.delete()
+    if request.method == 'POST':
+        obj = get_object_or_404(Task, srno=srno, user=request.user)
+        obj.delete()
     return redirect('/todopage/')
 
 @login_required(login_url='/loginn/')
@@ -29,6 +30,8 @@ def edit_todo(request, srno):
     if request.method == 'POST':
         title = request.POST.get('title')
         obj.title = title
+        deadline = request.POST.get('deadline')
+        obj.deadline = deadline
         obj.save()
         return redirect('/todopage/')
